+++ packages/ts-components/src/fixtures/preview-data/in-article-puff.ts
@@ -1,0 +1,41 @@
+export default {
+  41547: {
+    body: {
+      data: [
+        {
+          type: 'puff',
+          data: {
+            image: '',
+            label: 'Bricks & Mortar',
+            headline: 'How to build your own house',
+            copy:
+              'Prepare for the <b>Help to Build</b> equity loan scheme <i>before</i> it starts in the autumn',
+            link:
+              'https://www.thetimes.co.uk/article/how-to-build-your-own-house-a-beginners-guide-to-self-builds-v6khqxsbd',
+            linkText: ''
+          }
+        }
+      ]
+    }
+  },
+  41548: {
+    body: {
+      data: [
+        {
+          type: 'puff',
+          data: {
+            image:
+              'https://nuk-tnl-deck-prod-static.s3-eu-west-1.amazonaws.com/uploads/c1b6ab118e9422965b7faa628e26e05b.jpeg',
+            label: 'Bricks & Mortar',
+            headline: 'How to build your own house',
+            copy:
+              'Prepare for the <b>Help to Build</b> equity loan scheme <i>before</i> it starts in the autumn',
+            link:
+              'https://www.thetimes.co.uk/article/how-to-build-your-own-house-a-beginners-guide-to-self-builds-v6khqxsbd',
+            linkText: 'Read the article'
+          }
+        }
+      ]
+    }
+  }
+};