+++ packages/article/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.7.58](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.57...@times-components/article@7.7.58) (2021-08-23)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.7.57](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.56...@times-components/article@7.7.57) (2021-08-23)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.7.56](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.55...@times-components/article@7.7.56) (2021-08-23)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.7.55](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.54...@times-components/article@7.7.55) (2021-08-23)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.7.54](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.53...@times-components/article@7.7.54) (2021-08-20)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.7.53](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.52...@times-components/article@7.7.53) (2021-08-19)
 
 **Note:** Version bump only for package @times-components/article